build: first round of work to unify device lists
This modifies the build system to get lcec_conf to link to all of the
drivers.  This will make it possible to get rid of `slaveTypes[]` in
`lcec_conf.c`, once we've dealt with all of the slave*Params.

Issue #7
